backend/requirements.txt
# Core packages
Django==4.2.10
django-cors-headers==4.3.1
djangorestframework==3.14.0
python-dotenv==1.0.0
PyJWT==2.8.0
djangorestframework-simplejwt==5.3.1
whitenoise==6.6.0
django-filter==24.1

# Database
# psycopg2-binary will be installed separately
dj-database-url==2.1.0

# Documentation
drf-spectacular==0.27.0

# Image processing
# Pillow will be installed separately

# Utilities
requests==2.31.0
markdown==3.5.2
django-cleanup==9.0.0

# Deployment
gunicorn==21.2.0
